+++ rpc/backend/call_tx.go
@@ -183,7 +183,10 @@ func (b *Backend) SetTxDefaults(args evmtypes.TransactionArgs) (evmtypes.Transac
 		return args, errors.New("both gasPrice and (maxFeePerGas or maxPriorityFeePerGas) specified")
 	}
 
-	head := b.CurrentHeader()
+	head, err := b.CurrentHeader()
+	if err != nil {
+		return args, err
+	}
 	if head == nil {
 		return args, errors.New("latest header is nil")
 	}
@@ -425,7 +428,11 @@ func (b *Backend) GasPrice() (*hexutil.Big, error) {
 		result *big.Int
 		err    error
 	)
-	if head := b.CurrentHeader(); head.BaseFee != nil {
+	head, err := b.CurrentHeader()
+	if err != nil {
+		return nil, err
+	}
+	if head.BaseFee != nil {
 		result, err = b.SuggestGasTipCap(head.BaseFee)
 		if err != nil {
 			return nil, err